feat(aci): Move project/environment fields from the header into the body
