+++ static/app/views/detectors/components/forms/metric/metric.tsx
@@ -36,6 +36,7 @@ import {useIsMigratedExtrapolation} from 'sentry/views/detectors/components/deta
 import {AutomateSection} from 'sentry/views/detectors/components/forms/automateSection';
 import {AssignSection} from 'sentry/views/detectors/components/forms/common/assignSection';
 import {DescribeSection} from 'sentry/views/detectors/components/forms/common/describeSection';
+import {ProjectEnvironmentSection} from 'sentry/views/detectors/components/forms/common/projectEnvironmentSection';
 import {EditDetectorLayout} from 'sentry/views/detectors/components/forms/editDetectorLayout';
 import type {MetricDetectorFormData} from 'sentry/views/detectors/components/forms/metric/metricFormData';
 import {
@@ -73,6 +74,7 @@ function MetricDetectorForm() {
     <Stack gap="2xl" maxWidth={theme.breakpoints.xl}>
       <TransactionsDatasetWarningListener />
       <MigratedAlertWarningListener />
+      <ProjectEnvironmentSection />
       <TemplateSection />
       <CustomizeMetricSection />
       <DetectSection />
