+++ static/app/views/detectors/components/forms/uptime/index.tsx
@@ -13,8 +13,11 @@ import {useUptimeAssertionFeatures} from 'sentry/views/alerts/rules/uptime/useUp
 import {AutomateSection} from 'sentry/views/detectors/components/forms/automateSection';
 import {AssignSection} from 'sentry/views/detectors/components/forms/common/assignSection';
 import {DescribeSection} from 'sentry/views/detectors/components/forms/common/describeSection';
+import {
+  ProjectEnvironmentSection,
+  type EnvironmentConfig,
+} from 'sentry/views/detectors/components/forms/common/projectEnvironmentSection';
 import {useSetAutomaticName} from 'sentry/views/detectors/components/forms/common/useSetAutomaticName';
-import type {DetectorBaseFields} from 'sentry/views/detectors/components/forms/detectorBaseFields';
 import {EditDetectorLayout} from 'sentry/views/detectors/components/forms/editDetectorLayout';
 import {NewDetectorLayout} from 'sentry/views/detectors/components/forms/newDetectorLayout';
 import {ConnectedTestUptimeMonitorButton} from 'sentry/views/detectors/components/forms/uptime/connectedTestUptimeMonitorButton';
@@ -28,8 +31,10 @@ import {UptimeRegionWarning} from 'sentry/views/detectors/components/forms/uptim
 import {UptimeDetectorResolveSection} from 'sentry/views/detectors/components/forms/uptime/resolve';
 import {UptimeDetectorVerificationSection} from 'sentry/views/detectors/components/forms/uptime/verification';
 
-const ENVIRONMENT_CONFIG: React.ComponentProps<typeof DetectorBaseFields>['environment'] =
-  {includeAllEnvironments: false, fieldProps: {required: true}};
+const ENVIRONMENT_CONFIG: EnvironmentConfig = {
+  includeAllEnvironments: false,
+  fieldProps: {required: true},
+};
 
 function UptimeDetectorForm() {
   const theme = useTheme();
@@ -56,6 +61,7 @@ function UptimeDetectorForm() {
     <Stack gap="2xl" maxWidth={theme.breakpoints.lg}>
       <UptimeRegionWarning />
       <PreviewSection />
+      <ProjectEnvironmentSection environment={ENVIRONMENT_CONFIG} />
       <UptimeDetectorFormDetectSection />
       <UptimeDetectorVerificationSection />
       <UptimeDetectorResolveSection />
@@ -83,7 +89,6 @@ function NewUptimeDetectorFormContent() {
       detectorType="uptime_domain_failure"
       formDataToEndpointPayload={uptimeFormDataToEndpointPayload}
       initialFormData={{}}
-      environment={ENVIRONMENT_CONFIG}
       extraFooterButton={
         hasRuntimeAssertions ? <ConnectedTestUptimeMonitorButton /> : undefined
       }
@@ -111,7 +116,6 @@ function EditExistingUptimeDetectorFormContent({detector}: {detector: UptimeDete
       detector={detector}
       formDataToEndpointPayload={uptimeFormDataToEndpointPayload}
       savedDetectorToFormData={uptimeSavedDetectorToFormData}
-      environment={ENVIRONMENT_CONFIG}
       extraFooterButton={
         hasRuntimeAssertions ? <ConnectedTestUptimeMonitorButton size="sm" /> : undefined
       }
